New infra package to share code across the mono repo
And migrate last test in _test_yaml to null-safety
