+++ _test_yaml/pubspec.yaml
@@ -5,6 +5,8 @@ environment:
   sdk: '>=2.14.0 <3.0.0'
 
 dev_dependencies:
+  _json_serial_shared_test:
+    path: ../shared_test
   build_runner: ^2.0.0
   build_verify: ^3.0.0
   checked_yaml: any
