+++ _test_yaml/test/src/build_config.dart
@@ -12,7 +12,7 @@ class Config {
   final Map<String, Builder> builders;
 
   // Verifying enum keys in map
-  Map<AutoApply, int>? weights;
+  Map<AutoApply, int?>? weights;
 
   Config({required this.builders});
 
@@ -29,7 +29,7 @@ class Config {
 class Builder {
   final String? target;
 
-  final String import;
+  final String? import;
 
   @JsonKey(name: 'is_optional')
   final bool? isOptional;
@@ -58,7 +58,7 @@ class Builder {
   final Map<String, List<String>>? buildExtensions;
 
   Builder({
-    required this.import,
+    this.import,
     this.target,
     this.isOptional,
     this.autoApply,
