+++ json_serializable/pubspec.yaml
@@ -25,12 +25,13 @@ dependencies:
   source_helper: ^1.3.0
 
 dev_dependencies:
+  _json_serial_shared_test:
+    path: ../shared_test
   build_runner: ^2.0.0
   build_verify: ^3.0.0
   dart_style: ^2.0.0
   logging: ^1.0.0
   source_gen_test: ^1.0.0
-  stack_trace: ^1.10.0
   test: ^1.16.0
   test_descriptor: ^2.0.0
   test_process: ^2.0.0
