+++ json_serializable/test/generic_files/generic_argument_factories_nullable.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// @dart=2.12
-
 import 'package:json_annotation/json_annotation.dart';
 
 part 'generic_argument_factories_nullable.g.dart';
