+++ json_serializable/test/generic_files/generic_test.dart
@@ -318,7 +318,7 @@ void main() {
   });
 
   test('issue 980 regression test', () {
-    validateRoundTrip(
+    roundTripObject(
       Issue980ParentClass([
         Issue980GenericClass(45),
         Issue980GenericClass(42),
