+++ json_serializable/test/test_utils.dart
@@ -2,11 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:convert';
-
 import 'package:collection/collection.dart';
-import 'package:stack_trace/stack_trace.dart';
 import 'package:test/test.dart';
+
+export 'package:_json_serial_shared_test/shared_test.dart';
 
 final throwsTypeError = throwsA(isTypeError);
 
@@ -14,48 +13,3 @@ final isTypeError = isA<TypeError>();
 
 bool deepEquals(dynamic a, dynamic b) =>
     const DeepCollectionEquality().equals(a, b);
-
-void validateRoundTrip<T>(
-  T object,
-  T Function(Map<String, dynamic> json) factory,
-) {
-  final data = loudEncode(object);
-
-  final object2 = factory(json.decode(data) as Map<String, dynamic>);
-
-  expect(object2, equals(object));
-
-  final json2 = loudEncode(object2);
-
-  expect(json2, equals(data));
-}
-
-/// Prints out nested causes before throwing `JsonUnsupportedObjectError`.
-String loudEncode(Object? object) {
-  try {
-    return const JsonEncoder.withIndent(' ').convert(object);
-  } catch (e) {
-    if (e is JsonUnsupportedObjectError) {
-      Object? error = e;
-
-      var count = 1;
-
-      while (error is JsonUnsupportedObjectError) {
-        print(
-          '(${count++}) $error ${error.unsupportedObject} '
-          '(${error.unsupportedObject.runtimeType}) !!!',
-        );
-        print(Trace.from(error.stackTrace!).terse);
-        error = error.cause;
-      }
-
-      if (error != null) {
-        print('(${count++}) $error ???');
-        if (error is Error) {
-          print(Trace.from(error.stackTrace!).terse);
-        }
-      }
-    }
-    rethrow;
-  }
-}
