+++ shared_test/pubspec.yaml
@@ -1,0 +1,8 @@
+name: _json_serial_shared_test
+publish_to: none
+environment:
+  sdk: '>=2.14.0 <3.0.0'
+
+dependencies:
+  stack_trace: ^1.10.0
+  test: ^1.6.0
